requirements.txt
black==24.3.0
Django==4.1.10
django-environ==0.9.0
django-extensions==3.1.3
djangorestframework==3.14.0
drf-yasg==1.21.7
ipython==8.10.0
psycopg2==2.9.5
pytest==8.1.1
pytest-django==4.8.0